chore: Abstract styles out to their own files
