+++ components/AppHeader.tsx
@@ -3,36 +3,7 @@ import { AppBar, Box, Button, Chip, Container, IconButton, Menu, MenuItem, Toolb
 import Link from 'next/link'
 import { useState } from 'react'
 import ConnectedAccount from './ConnectedAccount'
-
-const styles = {
-	appbar: {
-		backgroundColor: '#fff',
-		color: '#111',
-		boxShadow: 0,
-	},
-	alphaChip: {
-		backgroundColor: '#FF5200',
-		color: '#fff',
-		ml: 1,
-		fontSize: '.6rem',
-		height: '1.25rem',
-		textShadow: 'none',
-	},
-	logoDesktop: {
-		mr: 2,
-		display: { xs: 'none', md: 'flex' },
-		cursor: 'pointer',
-		textTransform: 'uppercase',
-		fontSize: '2rem',
-	},
-	logoMobile: {
-		display: { xs: 'flex', alignItems: 'center', md: 'none' },
-		ml: 1,
-		cursor: 'pointer',
-		textTransform: 'uppercase',
-		fontSize: '2rem',
-	},
-}
+import styles from './AppHeader.styles'
 
 const pages = [
 	{ href: '/projects/new', title: 'Create' },
