+++ components/ConnectedAccount.styles.ts
@@ -1,0 +1,23 @@
+const styles = {
+	wrapper: {
+		display: 'flex',
+		justifyContent: 'flex-end',
+		alignItems: 'center',
+	},
+	btn: {
+		borderColor: '#000',
+		borderWidth: '3px',
+		fontWeight: 600,
+		fontStyle: 'italic',
+		'&:hover': {
+			borderWidth: '3px',
+		},
+	},
+	avatar: {
+		mr: 1,
+		p: 0,
+		border: '2px solid #000',
+	},
+}
+
+export default styles
