+++ components/ConnectedAccount.tsx
@@ -3,28 +3,7 @@ import Link from 'next/link'
 import { useState } from 'react'
 import formatAddress from '../utils/formatAddress'
 import { useWeb3 } from './Web3Provider'
-
-const styles = {
-	wrapper: {
-		display: 'flex',
-		justifyContent: 'flex-end',
-		alignItems: 'center',
-	},
-	btn: {
-		borderColor: '#000',
-		borderWidth: '3px',
-		fontWeight: 600,
-		fontStyle: 'italic',
-		'&:hover': {
-			borderWidth: '3px',
-		},
-	},
-	avatar: {
-		mr: 1,
-		p: 0,
-		border: '2px solid #000',
-	},
-}
+import styles from './ConnectedAccount.styles'
 
 const ConnectedAccount = (): JSX.Element => {
 	const [anchorEl, setAnchorEl] = useState(null)
