+++ components/NFTCard.tsx
@@ -5,60 +5,7 @@ import PolygonIcon from '../public/polygon_logo_black.png'
 import formatAddress from '../utils/formatAddress'
 import formatDate from '../utils/formatDate'
 import ImageOptimized from './ImageOptimized'
-
-const styles = {
-	buyableChip: {
-		textTransform: 'uppercase',
-		fontWeight: 800,
-		fontSize: '1rem',
-		position: 'absolute',
-		top: '1.5rem',
-		right: '-1rem',
-		backgroundColor: '#ff5200',
-		py: 2.5,
-		color: '#fff',
-		zIndex: 1,
-	},
-	card: {
-		minWidth: '200px',
-		position: 'relative',
-		overflow: 'visible',
-	},
-	projectIconLink: {
-		ml: 1,
-	},
-	createdAt: {
-		fontWeight: 300,
-		fontStyle: 'italic',
-		my: 1,
-	},
-	detailItem: {
-		textTransform: 'uppercase',
-		color: '#a8a8a8',
-		fontWeight: 600,
-		mb: 1,
-	},
-	actions: {
-		alignItems: 'flex-end',
-		justifyContent: 'space-between',
-	},
-	buyNowBtn: {
-		ml: 1,
-	},
-	buyNowPrice: {
-		mt: 3,
-		display: 'flex',
-		alignItems: 'center',
-	},
-	price: {
-		display: 'flex',
-		alignItems: 'center',
-	},
-	eth: {
-		color: '#aaa',
-		fontSize: '1rem',
-	},
-}
+import styles from './NFTCard.styles'
 
 type NFTCardProps = {
 	details: any // cid and token
