+++ components/StemCard.styles.ts
@@ -1,0 +1,27 @@
+const styles = {
+	card: {
+		minWidth: '200px',
+	},
+	cardMedia: {
+		borderTopLeftRadius: '4px',
+		borderTopRightRadius: '4px',
+		borderBottom: '3px solid #000',
+		py: 1,
+		px: 1.5,
+	},
+	cardMediaIcon: {
+		color: '#000',
+		fontSize: '2rem',
+	},
+	detailItem: {
+		textTransform: 'uppercase',
+		color: '#a8a8a8',
+		fontWeight: 600,
+		mb: 1,
+	},
+	actions: {
+		justifyContent: 'flex-end',
+	},
+}
+
+export default styles
