+++ components/StemCard.tsx
@@ -4,32 +4,7 @@ import Link from 'next/link'
 import PropTypes from 'prop-types'
 import formatAddress from '../utils/formatAddress'
 import formatStemName from '../utils/formatStemName'
-
-const styles = {
-	card: {
-		minWidth: '200px',
-	},
-	cardMedia: {
-		borderTopLeftRadius: '4px',
-		borderTopRightRadius: '4px',
-		borderBottom: '3px solid #000',
-		py: 1,
-		px: 1.5,
-	},
-	cardMediaIcon: {
-		color: '#000',
-		fontSize: '2rem',
-	},
-	detailItem: {
-		textTransform: 'uppercase',
-		color: '#a8a8a8',
-		fontWeight: 600,
-		mb: 1,
-	},
-	actions: {
-		justifyContent: 'flex-end',
-	},
-}
+import styles from './StemCard.styles'
 
 const propTypes = {
 	details: PropTypes.shape({
