+++ components/StemDropzone.styles.ts
@@ -1,0 +1,31 @@
+const styles = {
+	spinner: {
+		color: '#555',
+		my: 4,
+	},
+	uploadTitle: {
+		textAlign: 'center',
+	},
+	uploadIcon: {
+		fontSize: '4rem',
+		mb: 2,
+	},
+	uploadText: {
+		textAlign: 'center',
+		fontSize: '1.25rem',
+		fontWeight: 600,
+		mb: 2,
+	},
+	uploadMeta: {
+		textAlign: 'center',
+		fontStyle: 'italic',
+		fontSize: '10px',
+		mx: 'auto',
+	},
+	errorMsg: {
+		textAlign: 'center',
+		mt: 2,
+	},
+}
+
+export default styles
