+++ components/StemDropzone.tsx
@@ -2,36 +2,7 @@ import { Check, CloudUpload } from '@mui/icons-material'
 import { Box, Typography } from '@mui/material'
 import { useMemo } from 'react'
 import { useDropzone } from 'react-dropzone'
-
-const styles = {
-	spinner: {
-		color: '#555',
-		my: 4,
-	},
-	uploadTitle: {
-		textAlign: 'center',
-	},
-	uploadIcon: {
-		fontSize: '4rem',
-		mb: 2,
-	},
-	uploadText: {
-		textAlign: 'center',
-		fontSize: '1.25rem',
-		fontWeight: 600,
-		mb: 2,
-	},
-	uploadMeta: {
-		textAlign: 'center',
-		fontStyle: 'italic',
-		fontSize: '10px',
-		mx: 'auto',
-	},
-	errorMsg: {
-		textAlign: 'center',
-		mt: 2,
-	},
-}
+import styles from './StemDropzone.styles'
 
 const baseStyle = {
 	flex: 1,
