+++ pages/nfts/[id].tsx
@@ -16,92 +16,7 @@ import PolygonIcon from '../../public/polygon_logo_black.png'
 import formatAddress from '../../utils/formatAddress'
 import formatDate from '../../utils/formatDate'
 import { get, update } from '../../utils/http'
-
-const styles = {
-	title: {
-		textTransform: 'uppercase',
-		fontStyle: 'italic',
-		fontWeight: 800,
-		display: 'flex',
-		alignItems: 'center',
-	},
-	buyableChip: {
-		ml: 1,
-		textTransform: 'uppercase',
-		fontWeight: 800,
-		fontSize: '1rem',
-		backgroundColor: '#ff5200',
-		color: '#fff',
-	},
-	metadata: {
-		my: 2,
-	},
-	metadataKey: {
-		mr: 1,
-		display: 'inline-block',
-		color: '#a8a8a8',
-	},
-	buyNowListing: {
-		mt: 3,
-		display: 'flex',
-		alignItems: 'center',
-	},
-	buyNowBtn: {
-		fontWeight: 800,
-		fontStyle: 'italic',
-		fontSize: '1rem',
-		letterSpacing: '.5px',
-		color: '#111',
-	},
-	price: {
-		display: 'flex',
-		alignItems: 'center',
-		pl: 3,
-	},
-	eth: {
-		color: '#aaa',
-		fontSize: '1rem',
-	},
-	btn: {
-		mb: 2,
-		display: 'block',
-		textAlign: 'center',
-	},
-	divider: {
-		my: 3,
-		borderColor: '#ccc',
-	},
-	sectionMeta: {
-		fontStyle: 'italic',
-		fontWeight: 300,
-		textTransform: 'uppercase',
-		color: '#777',
-		mb: 2,
-	},
-	sectionCount: {
-		fontWeight: 300,
-		color: '#777',
-		ml: 1,
-		fontSize: '1.5rem',
-	},
-	collaborator: {
-		my: 3,
-		display: 'flex',
-		alignItems: 'center',
-	},
-	collaboratorAddress: {
-		ml: 1,
-		fontStyle: 'italic',
-	},
-	noItemsMsg: {
-		textAlign: 'center',
-		marginY: 4,
-	},
-	error: {
-		textAlign: 'center',
-		marginY: 4,
-	},
-}
+import { detailsStyles as styles } from '../../styles/NFTs.styles'
 
 const propTypes = {
 	covalentData: PropTypes.shape({
