+++ pages/projects/index.tsx
@@ -6,32 +6,7 @@ import PropTypes from 'prop-types'
 import ProjectCard from '../../components/ProjectCard'
 import { IProjectDoc } from '../../models/project.model'
 import { get } from '../../utils/http'
-
-const styles = {
-	title: {
-		textTransform: 'uppercase',
-		fontStyle: 'italic',
-		fontWeight: 800,
-		display: 'flex',
-		justifyContent: 'center',
-		alignItems: 'center',
-		mb: 2,
-	},
-	subtitle: {
-		fontStyle: 'italic',
-		fontWeight: 300,
-		textAlign: 'center',
-		mb: 7,
-	},
-	noProjects: {
-		textAlign: 'center',
-	},
-	noProjectsMsg: {
-		fontSize: '1.5rem',
-		color: '#555',
-		mb: 3,
-	},
-}
+import { indexStyles as styles } from '../../styles/Projects.styles'
 
 const propTypes = {
 	data: PropTypes.arrayOf(
@@ -58,7 +33,8 @@ const ProjectsPage: NextPage<ProjectsPageProps> = props => {
 					</Typography>
 					<Container maxWidth="sm">
 						<Typography variant="h5" sx={styles.subtitle}>
-							Explore the ecosystem for unique music, collaborate and build with others, create something one of a kind.
+							Explore the ecosystem for unique music, collaborate and build with others, create
+							something one of a kind.
 						</Typography>
 					</Container>
 					{data.length > 0 ? (
