+++ pages/stats/index.tsx
@@ -4,35 +4,7 @@ import CovalentInsights from '../../components/CovalentInsights'
 import PropTypes from 'prop-types'
 import Head from 'next/head'
 import Link from 'next/link'
-
-const styles = {
-	title: {
-		textTransform: 'uppercase',
-		fontStyle: 'italic',
-		fontWeight: 800,
-		display: 'flex',
-		justifyContent: 'center',
-		alignItems: 'center',
-		mb: 2,
-	},
-	subtitle: {
-		fontStyle: 'italic',
-		fontWeight: 300,
-		textAlign: 'center',
-		mb: 4,
-	},
-	text: {
-		display: 'block',
-		textAlign: 'center',
-		m: 0,
-		lineHeight: '1.3rem',
-		fontSize: '.95rem',
-	},
-	divider: {
-		my: 3,
-		borderColor: '#ccc',
-	},
-}
+import styles from '../../styles/Stats.styles'
 
 const propTypes = {
 	data: PropTypes.shape({
