+++ styles/Stats.styles.ts
@@ -1,0 +1,30 @@
+const styles = {
+	title: {
+		textTransform: 'uppercase',
+		fontStyle: 'italic',
+		fontWeight: 800,
+		display: 'flex',
+		justifyContent: 'center',
+		alignItems: 'center',
+		mb: 2,
+	},
+	subtitle: {
+		fontStyle: 'italic',
+		fontWeight: 300,
+		textAlign: 'center',
+		mb: 4,
+	},
+	text: {
+		display: 'block',
+		textAlign: 'center',
+		m: 0,
+		lineHeight: '1.3rem',
+		fontSize: '.95rem',
+	},
+	divider: {
+		my: 3,
+		borderColor: '#ccc',
+	},
+}
+
+export default styles
